requirements-dev.txt
appdirs==1.4.4
certifi==2020.4.5.1
chardet==3.0.4
distlib==0.3.0
entrypoints==0.3
filelock==3.0.12
flake8==3.8.1
idna==2.9
mccabe==0.6.1
pipenv==2018.11.26
pycodestyle==2.6.0
pyflakes==2.2.0

# Deps do Pytest
py==1.8.1
more-itertools==8.3.0
attrs==19.3.0
pluggy==0.13.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
virtualenv==20.0.20
virtualenv-clone==0.5.4
#Coverage
coverage==5.1
pytest-cov==2.8.1
-r requirements.txt

